:art: Move ItemClickListener
