--- dEventer/app/src/main/java/es/achraf/deventer/interfaces/ItemClickListener.java
@@ -1,8 +1,0 @@
-package es.achraf.deventer.interfaces;
-
-import android.view.View;
-
-public interface ItemClickListener {
-
-    void onItemClick(View view, int pos);
-}
