+++ dEventer/app/src/main/java/es/achraf/deventer/view/fragments/OwnEventsFragment.java
@@ -24,7 +24,7 @@ import java.util.ArrayList;
 
 import de.hdodenhof.circleimageview.CircleImageView;
 import es.achraf.deventer.R;
-import es.achraf.deventer.interfaces.ItemClickListener;
+import es.achraf.deventer.view.ItemClickListener;
 import es.achraf.deventer.model.Event;
 import es.achraf.deventer.view.IView;
 import es.achraf.deventer.view.adapters.RecyclerViewEventAdapter;
